pyproject.toml
[tool.poetry]
authors = ["ehco1996 <zh19960202@gmail.com>"]
description = "shadowsocks build with asyncio , also support many user in one port"
include = [
  "LICENSE",
]
keywords = ["shadowsocsk", "asyncio", "grpc"]
license = "GPLV3"
name = "aioshadowsocks"
readme = "README.md"
version = "0.1.7"

packages = [
  {include = "shadowsocks"},
]

[tool.poetry.dependencies]
aiohttp = "==3.6.2"
bloom-filter = "==1.3"
fire = "==0.3.1"
grpclib = "==0.4.1"
hkdf = "==0.0.3"
prometheus-async = "==19.2.0"
prometheus-client = "==0.8.0"
protobuf = "==3.13.0"
python = ">=3.8.0,<4.0"
sentry-sdk = "==0.18.0"
cryptography = "^3.2.1"
httpx = "^0.16.1"
grpcio = "^1.35.0"
tortoise-orm = "^0.17.1"

[tool.poetry.dev-dependencies]
autoflake = ">=1.4"
black = ">=20.8b1"
grpcio-tools = "^1.30.0"
isort = "5.5.4"
pytest = "^6.2.0"

[tool.poetry.scripts]
aioss = 'shadowsocks.__main__:main'

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
